refactor: excluindo arquivos individuais das telas principal,final e sobre para unificar em um único arquivo.
