--- src/telaFinalizacao.h
@@ -1,6 +1,0 @@
-#ifndef TELA_FINALIZACAO_H
-#define TELA_FINALIZACAO_H
-
-void telaFinalizacao(void);
-
-#endif
