--- src/telaPrincipal.h
@@ -1,6 +1,0 @@
-#ifndef TELA_PRINCIPAL_H
-#define TELA_PRINCIPAL_H
-
-char telaPrincipal(void);
-
-#endif
