--- src/telaSobre.h
@@ -1,6 +1,0 @@
-#ifndef TELA_SOBRE_H
-#define TELA_SOBRE_H
-
-void telaSobre(void);
-
-#endif
