+++ src/ui/outrasTelas.h
@@ -1,0 +1,8 @@
+#ifndef OUTRASTELAS_H
+#define OUTRASTELAS_H
+
+char telaPrincipal(void);
+void telaFinalizacao(void);
+void telaSobre(void);
+
+#endif
