--- src/ui/telaPrincipal.c
@@ -1,25 +1,0 @@
-#include <stdio.h>
-
-char telaPrincipal(void) {
-
-    char op;
-
-    printf("===================\n");
-    printf("SIG-GYM\n");
-    printf("===================\n");
-    printf("(1) ALUNOS\n");
-    printf("(2) PLANOS\n");
-    printf("(3) MATRICULAS\n");
-    printf("(4) CHECK-INS\n");
-    printf("(5) SOBRE\n");
-    printf("(0) SAIR\n");
-
-    scanf("%c", &op);
-    getchar();
-
-    return op;
-    }
-
-
-
-
