--- src/ui/telaSobre.c
@@ -1,17 +1,0 @@
-#include <stdio.h>
-
-void telaSobre(void) {
-    
-    printf("==================\n");
-    printf("SOBRE SIG-GYM\n");
-    printf("==================\n");
-    printf("O sistema, quando finalizado, deve\n");
-    printf("apresentar funcioalidades básicas que\n");
-    printf("permitam o gerenciamento simples de\n");
-    printf("uma academia.\n");
-    printf("==================\n");
-    printf("Desenvolvido por Rikelme Nicolau & Flávia Juliana\n");
-   
-    getchar();
-}
-
